feat: end 화면 구현
