+++ app/src/main/java/com/yourssu/unscramble/presentation/end/EndFragment.kt
@@ -1,4 +1,4 @@
-package com.yourssu.unscramble.presentation
+package com.yourssu.unscramble.presentation.end
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,16 +6,18 @@ import android.view.View
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import com.yourssu.unscramble.R
-import com.yourssu.unscramble.databinding.FragmentStartBinding
+import com.yourssu.unscramble.databinding.FragmentEndBinding
 import com.yourssu.unscramble.util.base.BindFragment
 import dagger.hilt.android.AndroidEntryPoint
 
-@AndroidEntryPoint
-class StartFragment : BindFragment<FragmentStartBinding>() {
 
-    private val viewModel: StartViewModel by viewModels()
-    override fun setBinding(layoutInflater: LayoutInflater): FragmentStartBinding {
-        return FragmentStartBinding.inflate(layoutInflater)
+@AndroidEntryPoint
+class EndFragment : BindFragment<FragmentEndBinding>() {
+
+    private val viewModel: EndViewModel by viewModels()
+
+    override fun setBinding(layoutInflater: LayoutInflater): FragmentEndBinding {
+        return FragmentEndBinding.inflate(layoutInflater)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -24,7 +26,10 @@ class StartFragment : BindFragment<FragmentStartBinding>() {
         binding.viewModel = viewModel
         binding.lifecycleOwner = viewLifecycleOwner
 
-        binding.btnStart.setOnClickListener {
+        binding.btnHome.setOnClickListener{
+            findNavController().navigate(R.id.startFragment)
+        }
+        binding.btnRetry.setOnClickListener{
             findNavController().navigate(R.id.timerFragment)
         }
     }
