+++ app/src/main/java/com/yourssu/unscramble/presentation/start/StartFragment.kt
@@ -1,0 +1,31 @@
+package com.yourssu.unscramble.presentation.start
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import androidx.fragment.app.viewModels
+import androidx.navigation.fragment.findNavController
+import com.yourssu.unscramble.R
+import com.yourssu.unscramble.databinding.FragmentStartBinding
+import com.yourssu.unscramble.util.base.BindFragment
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class StartFragment : BindFragment<FragmentStartBinding>() {
+
+    private val viewModel: StartViewModel by viewModels()
+    override fun setBinding(layoutInflater: LayoutInflater): FragmentStartBinding {
+        return FragmentStartBinding.inflate(layoutInflater)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        binding.viewModel = viewModel
+        binding.lifecycleOwner = viewLifecycleOwner
+
+        binding.btnStart.setOnClickListener {
+            findNavController().navigate(R.id.timerFragment)
+        }
+    }
+}
