+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/startFragment"
-        android:name="com.yourssu.unscramble.presentation.StartFragment"
+        android:name="com.yourssu.unscramble.presentation.start.StartFragment"
         android:label="fragment_start"
         tools:layout="@layout/fragment_start" >
         <action
@@ -25,7 +25,7 @@
     </fragment>
     <fragment
         android:id="@+id/endFragment"
-        android:name="com.yourssu.unscramble.presentation.EndFragment"
+        android:name="com.yourssu.unscramble.presentation.end.EndFragment"
         android:label="fragment_end"
         tools:layout="@layout/fragment_end" />
     <fragment
